requirements.txt
dnspython
PyNaCl
timeout
pandas
discord.py
webhooks
discord-pretty-help
discord-py-slash-command
aiohttp~=3.7.4